+++ app/views/playground.vue
@@ -35,7 +35,7 @@
 			<div class="sheet-container" :class="{loading: engraving, dirty: engraverDirty, chromatic: chromaticSymbols}">
 				<SheetSimple v-if="svgDocuments && !tokenizeStaff" :documents="svgDocuments" />
 				<MidiRoll v-if="midiPlayer" v-show="rollVisible" :player="midiPlayer" :timeScale="4e-3" :height="120" />
-				<SheetLive v-if="tokenizeStaff && sheetDocument"
+				<SheetLive v-if="tokenizeStaff && sheetDocument" ref="sheet"
 					:doc="sheetDocument"
 					:hashTable="svgHashTable"
 					:midi="midi"