+++ inc/staffSvg/index.ts
@@ -7,11 +7,16 @@ import structureTokens from "./structureTokens";
 
 const parseSvgPage = (dom, options) => {
 	const elem = svgToElements(dom, options);
+	//console.log("elem:", elem);
 	const {tokens, hashTable} = tokenizeElements(elem.children);
 
+	const [x, y, width, height] = elem.viewBox.match(/[\d-.]+/g).map(Number);
+	const viewBox = {x, y, width, height};
+
 	return {
-		structure: structureTokens(tokens),
+		structure: structureTokens(tokens, viewBox),
 		hashTable,
+		
 	};
 };