+++ inc/staffSvg/structureTokens.ts
@@ -187,7 +187,7 @@ const parseTokenMeasure = (tokens, endX) => {
 };
 
 
-const structureTokens = (tokens, viewBox) => {
+const structureTokens = (tokens, {viewBox, width, height}: any = {}) => {
 	const meaningfulTokens = tokens.filter(token => !token.is("NULL"));
 	const rowTokens = tokensRowsSplit(meaningfulTokens);
 
@@ -196,8 +196,9 @@ const structureTokens = (tokens, viewBox) => {
 	return {
 		rows,
 
-		width: viewBox.width,
-		height: viewBox.height,
+		viewBox,
+		width,
+		height,
 	};
 };